Move FileLogHandler to a separate module
