+++ handler-file-log/build.gradle
@@ -1,0 +1,64 @@
+plugins {
+    id 'com.android.library'
+    id 'maven-publish'
+}
+
+android {
+    namespace 'pl.brightinventions.slf4android.handler.filelog'
+    compileSdk 34
+
+    defaultConfig {
+        minSdk 19
+
+        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
+        consumerProguardFiles "consumer-rules.pro"
+    }
+
+    buildTypes {
+        release {
+            minifyEnabled false
+            proguardFiles getDefaultProguardFile('proguard-android-optimize.txt'), 'proguard-rules.pro'
+        }
+    }
+
+    publishing {
+        singleVariant('release') {
+            withSourcesJar()
+        }
+    }
+
+    java {
+        toolchain {
+            languageVersion = JavaLanguageVersion.of(17)
+        }
+    }
+
+    compileOptions {
+        sourceCompatibility JavaVersion.VERSION_1_8
+        targetCompatibility JavaVersion.VERSION_1_8
+    }
+}
+
+dependencies {
+    api project(':core')
+
+    androidTestImplementation 'org.hamcrest:hamcrest-core:1.3'
+    androidTestImplementation 'org.hamcrest:hamcrest-library:1.3'
+    androidTestImplementation('junit:junit:4.13.2')
+    androidTestImplementation 'androidx.test:rules:1.5.0'
+    androidTestImplementation 'org.easytesting:fest-util:1.2.5'
+}
+
+publishing {
+    publications {
+        release(MavenPublication) {
+            groupId = 'com.github.bright.slf4android'
+            artifactId = 'slf4android-handler-file-log'
+            version = '1.0'
+
+            afterEvaluate {
+                from components.release
+            }
+        }
+    }
+}
