+++ handler-file-log/src/main/java/pl/brightinventions/slf4android/handler/filelog/FileLogHandlerConfiguration.java
@@ -1,8 +1,11 @@
-package pl.brightinventions.slf4android;
+package pl.brightinventions.slf4android.handler.filelog;
 
 import android.content.Context;
 
 import java.util.logging.Handler;
+
+import pl.brightinventions.slf4android.LogRecordFormatter;
+import pl.brightinventions.slf4android.LoggerConfiguration;
 
 public abstract class FileLogHandlerConfiguration extends Handler {
     @SuppressWarnings("UnusedDeclaration")
@@ -18,11 +21,11 @@ public abstract class FileLogHandlerConfiguration extends Handler {
 
 
     /**
-     * Creates a {@link pl.brightinventions.slf4android.FileLogHandlerConfiguration} logger handler.
+     * Creates a {@link FileLogHandlerConfiguration} logger handler.
      * The returned new instance should be added to some logger with
      * {@link pl.brightinventions.slf4android.LoggerConfiguration#addHandlerToLogger(String, Handler)}
      *
-     * @return A new instance of {@link pl.brightinventions.slf4android.FileLogHandlerConfiguration}.
+     * @return A new instance of {@link FileLogHandlerConfiguration}.
      */
     public static FileLogHandlerConfiguration create(Context context) {
         LogRecordFormatter formatter = LoggerConfiguration.configuration()
